rpc - refactor
