+++ lib/index.js
@@ -13,9 +13,9 @@ const mkdirp = require('mkdirp');
 const defaults = require('../defaults.json');
 const upnp = require('./upnp.js');
 const mining = require('./mine.js');
-const EthRPC = require('./ethRPC.js');
-const XhrRpc = require('./xhrRPC.js');
-const WsRpc = require('./wsRPC.js');
+const EthRPC = require('./rpc/index.js');
+const HttpRpc = require('./rpc/server/http.js');
+const WsRpc = require('./rpc/server/ws.js');
 const blockProcesser = require('./blockProcesser.js');
 var genesisAllotments = require('ethereum-common').allotments;
 const networking = require('./networking.js');
@@ -99,7 +99,7 @@ App.prototype.start = function(cb) {
 
     //account manager doesn't do anything yet
     // self.accountMan = new AccountMan(detailsDB);
-    self.rpc = new EthRPC({app: self});
+    self.rpc = new EthRPC(app);
 
     //start the blockchain. This will lookup last block on the blockchain.
     self.blockchain.init(done);
@@ -152,14 +152,14 @@ App.prototype.start = function(cb) {
   }
 
   // XHR RPC Server
-  function startXhrRpcServer(done) {
-    self.xhrRpc = new XhrRpc({ app: self });
+  function startHttpRpcServer(done) {
+    self.xhrRpc = new HttpRpc(self.rpc);
     self.xhrRpc.start(self.settings.rpc.xhr, done);
   }
 
   // WS RPC Server
   function startWsRpcServer(done) {
-    self.wsRpc = new WsRpc({app: self});
+    self.wsRpc = new WsRpc(self.rpc);
     self.wsRpc.start(self.settings.rpc.ws, done);
   }
 
@@ -186,7 +186,7 @@ App.prototype.start = function(cb) {
   }
 
   if (this.settings.rpc) {
-    if(this.settings.rpc.xhr) tasks.startXhrRpcServer = ['setup', startXhrRpcServer];
+    if(this.settings.rpc.xhr) tasks.startHttpRpcServer = ['setup', startHttpRpcServer];
     if(this.settings.rpc.ws) tasks.startWsRpcServer = ['setup', startWsRpcServer];
   }
