+++ lib/rpc/server/http.js
@@ -5,11 +5,9 @@ const async = require('async');
 
 module.exports = XhrRpc;
 
-function XhrRpc(opts){
+function XhrRpc(rpc){
 
-  var app = this.app = opts.app;
   var server = this.server = Express();
-  var rpc = this.rpc = opts.app.rpc;
 
   // parse application/json even if it doesnt say its json (*/* does not work)
   server.use(bodyParser.json({ type: 'application/*' }));
