+++ lib/rpc/server/ws.js
@@ -3,9 +3,8 @@ const Rpc = require('./ethRPC.js');
 
 module.exports = WsRpc;
 
-function WsRpc(opts){
-  this.app = opts.app;
-  this.rpc = opts.app.rpc;
+function WsRpc(rpc){
+  this.rpc = rpc;
 }
 
 WsRpc.prototype.start = function(options, done) {
